requirements.txt
flask==3.0.2
pillow==10.2.0
moviepy==1.0.3
google-generativeai==0.4.0
requests==2.31.0
python-dotenv==1.0.1
yt-dlp==2023.12.30